// packages/shared/src/generated/registry.ts
export const CONTRACTS: readonly ContractDef[] = [
  {
    "name": "DNS",
    "slug": "dns",
    "address": "5CQax2tWAwS78Rzua3dr2njrHDs3gSN6Xf4AS2rYidtrNWoH",
    "abiFile": "dns.json",
    "codeHash": "0x76a65679c4016487b8df100b5a21e92b2ce13328ed9f57105f886d5600f23a47"
  },
  {
    "name": "Exchange",
    "slug": "exchange",
    "address": "5H7s7P2sJxKa9zbXZDkXitRy5pd8CfMB7uCBKE8H2nruKRJD",
    "abiFile": "geode_private_exchange.json",
    "codeHash": "0xf37843b048c9e8e38100f747c7547a537b332706bf84f841b27edab6c1c07ed0"
  },
  {
    "name": "LifeAndWork",
    "slug": "lifeandwork",
    "address": "5FYNEXcseSEH9hbNwK9ZtrAMhfymuCZ6Gys1oSu4Bka9TafY",
    "abiFile": "life_and_work.json",
    "codeHash": "0x2efcac2f70cac4495cf1c7040342a98b4e66dab46796105c017c6a0aeeb4f109"
  },
  {
    "name": "Market",
    "slug": "market",
    "address": "5FkPNGBky2u5bw2f5hKs13vpockvGaRGeGhtNyjHauwWVtoL",
    "abiFile": "geode_marketplace.json",
    "codeHash": "0x1886c0f55a89b4a81ff70bd23823c21cf39252cd965172e02fbde3cd5c7a8a4c"
  },
  {
    "name": "Messaging",
    "slug": "messaging",
    "address": "5Gt3CXXe6pmKfkyYSW4qSHNHtaqxQ2gjmTgr8iDjFzF6vUkQ",
    "abiFile": "geode_messaging.json",
    "codeHash": "0xff35e445209afa03d68fb194b1bc7016bf018f8e774689f575cd88d5a120647c"
  },
  {
    "name": "Profile",
    "slug": "profile",
    "address": "5Grs6qz9gfrX7SDubco845D5uRZTRnuyme8B2aFVjv6pHiQx",
    "abiFile": "geode_profile.json",
    "codeHash": "0xc101d2128b8098a1e2a168ffdc7fcfc8b444d539ca28cc43c3b8afb614119fa9"
  },
  {
    "name": "Social",
    "slug": "social",
    "address": "5FbLpQ5oHkTGUHD3Z6z7qas7MNnHH31mXsaxYFhgPA6ew7X7",
    "abiFile": "geode_social.json",
    "codeHash": "0xf3d4f4216f1d78d589455b86276924c8fb7c3a4a3f78c89403ded156902f98ec"
  },
  {
    "name": "Surveys",
    "slug": "surveys",
    "address": "5Dq6G1f42A8MDf9FGnTqJjCmrpeLMC3YaGFvSWv11hPsmiRL",
    "abiFile": "surveys.json",
    "codeHash": "0x3c4210e70c9822109962402bc9506d89f5d202d5eb48d977709306b98d8f25ac"
  }
];